Class_Archive/599 Matrix Methods for ML/sound_samples/pyproject.toml
[tool.poetry]
name = "599-sound-samples"
version = "0.1.0"
description = ""
authors = ["Molly Jones <molly.jones@spantree.net>"]
license = "CC BY-NC-SA 4.0 DEED"
readme = "README.md"
packages = [{include = "599_sound_samples"}]

[tool.poetry.dependencies]
python = "^3.10"
freesound-python = {git = "https://github.com/MTG/freesound-python"}
dotenv = "^0.0.5"
python-dotenv = "^1.0.1"
dvc = "^3.44.0"
dagshub = "^0.3.13"
boto3 = "^1.34.41"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
